Tests/LibWeb/Text/expected/css/insert-rule-keyframes-starts-animation.txt

before: 1
after: 1
keyframes: 0
